Extract user message renderer into render_user.py
Move ThemeColors to watch/types.py to avoid circular imports.
